use np in Spacecraft fields

+++ src/satsim/spacecraft/__init__.py
@@ -1,13 +1,15 @@
 from dataclasses import dataclass
 
+import numpy as np
+
 from satsim.types import StateVector
 
 
 @dataclass
 class Spacecraft:
     name: str
-    mass: float  # kg
-    ballistic_coefficient: float  # kg/m^2
+    mass: np.float64  # kg
+    ballistic_coefficient: np.float64  # kg/m^2
     state: StateVector
     deployed: bool = False
     attached: bool = True
